MAINT: refactored app components into several modules
