+++ dtool_lookup_gui/.gitignore
@@ -1,0 +1,1 @@
+gschemas.compiled
